+++ packages/genesys/packages/react-native-components/src/feedback/Dialog/Dialog.styles.ts
@@ -1,7 +1,6 @@
 import styled from "@peersyst/react-native-styled";
 import { Text } from "react-native";
 import { Modal } from "../Modal";
-import { DialogOptionProps } from "@peersyst/react-components-core";
 
 export const DialogRoot = styled(Modal)(() => ({
     width: "90%",
@@ -19,17 +18,3 @@ export const DialogMessage = styled(Text)(({ theme }) => ({
     color: theme.palette.text,
     fontSize: 14,
 }));
-
-export const DialogOption = styled(Text)<DialogOptionProps>(({ theme, type = "default" }) => {
-    const color =
-        type === "default"
-            ? theme.palette.primary
-            : type === "destructive"
-            ? theme.palette.status.error
-            : theme.palette.status.success;
-    return {
-        color,
-        fontWeight: "bold",
-        fontSize: 16,
-    };
-});