+++ packages/genesys/packages/react-native-components/src/input/Button/hooks/useButtonStyles.ts
@@ -4,6 +4,7 @@ import { extractTextStyles } from "@peersyst/react-native-utils";
 import { ButtonStyle, ButtonStyles } from "../Button.types";
 import { useMemo } from "react";
 import { ButtonSize, ButtonVariant } from "@peersyst/react-components-core";
+import { emphasize, getLuminance } from "@peersyst/react-utils";
 
 export interface UseButtonStylesResult {
     textStyle: TextStyle;
@@ -16,6 +17,7 @@ export default function useButtonStyles(
     size: ButtonSize,
     disabled: boolean,
     pressed: boolean,
+    color: string | undefined,
 ): UseButtonStylesResult {
     const { defaultStyles, defaultDisabledStyles, defaultPressedStyles, defaultSizeStyles } =
         useDefaultStyles();
@@ -46,9 +48,15 @@ export default function useButtonStyles(
         [defaultPressedStyles, pressedStyles],
     );
 
+    const pressedBackgroundColor = color ? emphasize(color, 0.4) : undefined;
     const [variantPressedTextStyles, variantPressedRootStyles] = useMemo(() => {
         return extractTextStyles({
             ...defaultPressedStyles.variant?.[variant],
+            ...(color
+                ? {
+                      gradient: { colors: [pressedBackgroundColor!, pressedBackgroundColor!] },
+                  }
+                : {}),
             ...variantPressedStyles?.[variant],
         });
     }, [variantPressedStyles, defaultPressedStyles, variant]);
@@ -57,6 +65,12 @@ export default function useButtonStyles(
         () =>
             extractTextStyles({
                 ...defaultStyles.variant?.[variant],
+                ...(color
+                    ? {
+                          gradient: { colors: [color, color] },
+                          color: getLuminance(color) > 0.5 ? "#000" : "#FFF",
+                      }
+                    : {}),
                 ...styles.variant?.[variant],
             }),
         [defaultStyles, styles, variant],